requirements.txt
telethon==1.36.0
python-dotenv==1.0.1
cryptg==0.4.0        # optional but speeds up media decryption significantly
